AC792N_SDK_RELEASE_V2.2.0_2025-03-19

+++ sdk/include_lib/driver/device/video/pipeline/core/pipeline_port.h
@@ -10,7 +10,7 @@
 
 typedef struct pipeline_endpoint {
     struct list_head entry;
-    int channel;
+    long long channel;
     char name[10];
     buffer_t *data_buffer;
     void *priv_data;
